MVP code refactoring
