+++ src/GenerationTree/Main.java
@@ -1,5 +1,7 @@
-import GenerationTree.interfaces.View;
+package GenerationTree;
+
 import GenerationTree.ui.ConsoleUI;
+import GenerationTree.ui.View;
 
 public class Main {
     public static void main(String[] args) {
