+++ src/GenerationTree/Person/Structs/Gender.java
@@ -1,4 +1,4 @@
-package GenerationTree.Structs;
+package GenerationTree.Person.Structs;
 
 public enum Gender {
     MALE {
